feature: support token exchange (#109)
This adds a token exchange command which can be used to execute the token exchange flow. Supports all token exchange parameters as specified in RFC8693. The subject token is mandatory and can be provided either on the command line or through stdin, similar to how the refresh token command has been implemented previously.